requirements.txt
allure-pytest==2.13.2
pytest==7.3.1
playwright==1.34.0
python-slugify~=8.0.1
requests~=2.31.0